refactor(package-manager-deb): implement single responsibility pattern
- Split monolithic main.go into focused modules:
  - main.go: Clean entry point with plugin metadata
  - package_manager.go: Execute method and command routing
  - download.go: Secure .deb package download with validation
  - dependency.go: Dependency resolution and management
  - validation.go: Comprehensive input validation with security checks

- Enhanced functionality:
  - Robust .deb package download and validation
  - Dependency resolution with apt integration
  - Enhanced security with URL and package validation
  - Better error handling and user feedback

- Follows same modular pattern as package-manager-apt
- Each file maintains single responsibility
- Improved maintainability and security
